Implement initial support for no_std

+++ .github/workflows/build_static.sh
@@ -10,20 +10,25 @@ main() {
         release_flag="--release"
     fi
 
+    if [ "$NO_STD" = "true" ]; then
+        (cd crates/no-std-test && cargo build --target $TARGET $FEATURES)
+        return
+    fi
+
     case $TARGET in
         asmjs-unknown-emscripten)
-            $cargo build -p livesplit --target $TARGET --release
+            $cargo build -p livesplit --target $TARGET --release $FEATURES
             ;;
         wasm32-unknown-emscripten)
             rm target/wasm32-unknown-emscripten/release/deps/*.wasm 2>/dev/null || :
             rm target/wasm32-unknown-emscripten/release/deps/*.js 2>/dev/null || :
-            $cargo build -p livesplit --target $TARGET --release
+            $cargo build -p livesplit --target $TARGET --release $FEATURES
             ;;
         wasm32-unknown-unknown)
-            $cargo build -p cdylib --target $TARGET --release
+            $cargo build -p cdylib --target $TARGET --release $FEATURES
             ;;
         *)
-            $cargo build -p staticlib --target $TARGET $release_flag
+            $cargo build -p staticlib --target $TARGET $release_flag $FEATURES
             ;;
     esac
 }
